Create Article Template
Template for articles from the weekly wellness update to be dropped in to. 

(Hey Zen, this is Oliver. I'm learning the basics of html so I'll be able to write and post articles with the help of chat gbt. if posible I'd love to have a meeting to touch base and see if that's possible. 
All the best, Oliver.)